:sparkles: Namespaces: Project Item
